+++ CHANGELOG.md
@@ -5,7 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2024-08-20
+
+### Added
+- Basic project structure and initial implementation of core components
+- `exchange` module with:
+    - Generic `Exchange` trait
+    - `OrderBook` structure for managing bids and asks
+    - `ExchangeUpdate` structure for handling updates
+- `bybit` module with initial `BybitExchange` implementation
+- `arbitrage_manager` module with `ArbitrageManager` structure for managing exchanges and analyzing opportunities
+
 ## [0.0.1] - 2024-08-20
 
-#### Added
+### Added
 - Initial repo files